requirements.txt
streamlit
requests
pandas
numpy
scikit-learn
nltk
python-dotenv
